ai-matcher/requirements.txt
flask
sentence-transformers
torch
PyPDF2
python-docx